application_files/xgboost/requirements.txt
#Flask
scikit-learn 
pandas      
numpy
pillow 
Flask>=2.1
Werkzeug==2.0.3

